Added test for builder resolution
