+++ src/DISCOSweb-Sdk/DISCOSweb-Sdk.Tests/Clients/DiscosClientTests.cs
@@ -1,20 +1,12 @@
 using System;
 using System.Collections;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Reflection;
 using System.Threading.Tasks;
 using DISCOSweb_Sdk.Clients;
-using DISCOSweb_Sdk.Models.ResponseModels;
-using DISCOSweb_Sdk.Models.ResponseModels.DiscosObjects;
 using DISCOSweb_Sdk.Models.ResponseModels.Entities;
-using DISCOSweb_Sdk.Models.ResponseModels.FragmentationEvent;
-using DISCOSweb_Sdk.Models.ResponseModels.Launches;
-using DISCOSweb_Sdk.Models.ResponseModels.LaunchVehicles;
-using DISCOSweb_Sdk.Models.ResponseModels.Orbits;
-using DISCOSweb_Sdk.Models.ResponseModels.Propellants;
-using DISCOSweb_Sdk.Models.ResponseModels.Reentries;
+using DISCOSweb_Sdk.Tests.TestDataGenerators;
 using Shouldly;
 using Xunit;
 
@@ -66,39 +58,4 @@ public class DiscosClientTests
 
 		((IEnumerable)resProperty!.GetValue(resTask)!).Cast<object>().Count().ShouldBeGreaterThan(1);
 	}
-}
-
-
-public class DiscosModelTypesTestDataGenerator : IEnumerable<object[]>
-{
-	private readonly Dictionary<Type, string> _ids = new()
-													 {
-														 {typeof(DiscosObject), "1"},
-														 {typeof(DiscosObjectClass), "6a6527d6-efbb-5500-abe3-594ac23d04ed"},
-														 {typeof(Country), "258"},
-														 {typeof(Organisation), "1699"},
-														 {typeof(Entity), "1"},
-														 {typeof(FragmentationEvent), "86"},
-														 {typeof(Launch), "1"},
-														 {typeof(LaunchSite), "14"},
-														 {typeof(LaunchSystem), "13"},
-														 {typeof(LaunchVehicle), "207"},
-														 {typeof(LaunchVehicleFamily), "50"},
-														 {typeof(LaunchVehicleEngine), "87381"},
-														 {typeof(LaunchVehicleStage), "328"},
-														 {typeof(InitialOrbitDetails), "1255"},
-														 {typeof(DestinationOrbitDetails), "36260"},
-														 {typeof(Propellant), "1"},
-														 {typeof(Reentry), "29338"}
-													 };
-
-	public IEnumerator<object[]> GetEnumerator()
-	{
-		return typeof(DiscosObject).Assembly
-								   .GetTypes()
-								   .Where(t => t.IsAssignableTo(typeof(DiscosModelBase)) && t != typeof(DiscosModelBase) && t != typeof(OrbitDetails))
-								   .Select(o => new object[] {o, _ids[o]}).GetEnumerator();
-	}
-
-	IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 }
